promote download on stats page
